+++ include/lexer.h
@@ -32,4 +32,7 @@ typedef struct s_redirection {
 
 } t_redirection;
 
+char	*ft_strndup(char const *s1, size_t n);
+t_list	*tokenizer(char *input);
+
 #endif
