+++ src/utils.c
@@ -1,0 +1,19 @@
+# include <stdlib.h>
+
+char	*ft_strndup(char const *s1, size_t n)
+{
+	size_t	i;
+	char	*data;
+
+	data = (char *)malloc(sizeof(char) * (n + 1));
+	if (!data)
+		return (0);
+	i = 0;
+	while (i < n)
+	{
+		data[i] = s1[i];
+		i++;
+	}
+	data[i] = '\0';
+	return (data);
+}
